pyproject.toml
[project]
name = "livekit-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "jex jang",email = "jexjang92@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector] (==1.0.22)",
    "mem0ai (>=0.1.102,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "livekit_demo.main:main"